requirements.txt
mysql-connector-python
pandas
PyPDF2
groq
python-dotenv
streamlit
SpeechRecognition
pyaudio
